refactor: asignación de la lógica a ComplaintService.java y CompanyService
