+++ src/main/java/com/uptc/edu/main/controller/ComplaintController.java
@@ -13,10 +13,8 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import com.uptc.edu.main.model.Company;
-import com.uptc.edu.main.model.Complaint;
 import com.uptc.edu.main.service.CompanyService;
 import com.uptc.edu.main.service.ComplaintService;
-import com.uptc.edu.main.service.SendEmail;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpSession;
@@ -30,13 +28,6 @@ public class ComplaintController {
     @Autowired
     private CompanyService companyService;
 
-    @Autowired
-    private final SendEmail sendEmail;
-
-    public ComplaintController(SendEmail sendEmail) {
-        this.sendEmail = sendEmail;
-    }
-
     @GetMapping("/registro")
     public String showForm(Model model) {
         model.addAttribute("entidades", getCompanyNames());
@@ -49,15 +40,7 @@ public class ComplaintController {
             @RequestParam String descripcion,
             Model model) {
 
-        companyService.searchByName(companyName).ifPresentOrElse(company -> {
-            Complaint complaint = new Complaint();
-            complaint.setDescription(descripcion);
-            complaint.setCompany(company);
-            complaintService.saveComplaint(complaint);
-            addMessage(model, "La queja fue registrada exitosamente.", "success");
-        }, () -> {
-            addMessage(model, "Error: empresa no encontrada.", "error");
-        });
+        companyService.createComplaintForExistingCompany(companyName, descripcion, model);
 
         model.addAttribute("entidades", getCompanyNames());
         return "registro";
@@ -70,22 +53,11 @@ public class ComplaintController {
                 .toList();
     }
 
-    private void addMessage(Model model, String message, String type) {
-        model.addAttribute("mensaje", message);
-        model.addAttribute("tipoMensaje", type);
-    }
-
     @GetMapping("/quejas")
     public String showComplaintsByCompany(@RequestParam(required = false) Long companyId, Model model) {
         model.addAttribute("entidades", companyService.listCompanies());
-        model.addAttribute("quejas", obtainVisibleComplaints(companyId));
+        model.addAttribute("quejas", complaintService.obtainVisibleComplaints(companyId));
         return "buscar";
-    }
-
-    private List<Complaint> obtainVisibleComplaints(Long companyId) {
-        return (companyId == null)
-                ? complaintService.findByIsVisibleTrue()
-                : complaintService.findByCompanyIdAndIsVisibleTrue(companyId);
     }
 
     @PatchMapping("/quejas/{id}/ocultar")
@@ -94,14 +66,7 @@ public class ComplaintController {
             RedirectAttributes redirectAttributes,
             HttpSession session) {
 
-        complaintService.searchById(id).ifPresentOrElse(complaint -> {
-            complaint.setVisible(false);
-            complaintService.saveComplaint(complaint);
-            redirectAttributes.addFlashAttribute("mensaje", "Queja eliminada exitosamente");
-            session.setAttribute("ultimaEmpresaBuscada", complaint.getCompany().getId());
-        }, () -> {
-            redirectAttributes.addFlashAttribute("error", "La queja no existe");
-        });
+        complaintService.hideComplaintIfExists(id, redirectAttributes, session);
 
         Long companyId = (Long) session.getAttribute("ultimaEmpresaBuscada");
         return "redirect:/quejas" + (companyId != null ? "?companyId=" + companyId : "");
@@ -115,17 +80,8 @@ public class ComplaintController {
 
         model.addAttribute("entidades", companyService.findAll());
 
-        companyService.searchById(entidadId).ifPresentOrElse(company -> {
-            List<Complaint> complaint = complaintService.findByCompanyIdAndIsVisibleTrue(company.getId());
-            model.addAttribute("quejas", complaint);
-            model.addAttribute("entidadSeleccionada", company.getName());            
-            sendEmail.sendEmail(request);
-        }, () -> {
-            model.addAttribute("quejas", List.of());
-            model.addAttribute("entidadSeleccionada", "Entidad no encontrada");
-        });
+        companyService.getCompanyComplaintsAndSendNotification(entidadId, model, request);
 
         return "buscar";
     }
-
 }
