+++ src/main/java/com/uptc/edu/main/service/ComplaintService.java
@@ -5,10 +5,13 @@ import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import com.uptc.edu.main.model.Company;
 import com.uptc.edu.main.model.Complaint;
 import com.uptc.edu.main.repository.ComplaintRepo;
+
+import jakarta.servlet.http.HttpSession;
 
 @Service
 public class ComplaintService {
@@ -38,5 +41,22 @@ public class ComplaintService {
 
     public List<Complaint> findByCompanyIdAndIsVisibleTrue(Long companyId) {
         return complaintRepo.findByCompanyIdAndIsVisibleTrue(companyId);
+    }
+
+    public List<Complaint> obtainVisibleComplaints(Long companyId) {
+        return (companyId == null)
+                ? complaintRepo.findByIsVisibleTrue()
+                : complaintRepo.findByCompanyIdAndIsVisibleTrue(companyId);
+    }
+
+    public void hideComplaintIfExists(Long id, RedirectAttributes redirectAttributes, HttpSession session) {
+        complaintRepo.findById(id).ifPresentOrElse(complaint -> {
+            complaint.setVisible(false);
+            complaintRepo.save(complaint);
+            redirectAttributes.addFlashAttribute("mensaje", "Queja eliminada exitosamente");
+            session.setAttribute("ultimaEmpresaBuscada", complaint.getCompany().getId());
+        }, () -> {
+            redirectAttributes.addFlashAttribute("error", "La queja no existe");
+        });
     }
 }
